Introduce recent_cpu value in threads
Threads have a recent_cpu value. This value is initialized to 0 for the
first thread. New threads inherit the recent_cpu of their parents.

If a non-idle thread is running, its recent_cpu value is incremented by
1 for every timer tick.

Every second, thread_update_recent_cpus() is called to update the recent
cpu values for each thread. However, this function is currently a no-op.